+++ doc/sphinx/source/releasenotes.rst
@@ -18,6 +18,7 @@ New features
 ^^^^^^^^^^^^
 * New HIP MAGMA backends for hipMAGMA library users: ``/gpu/hip/magma`` and ``/gpu/hip/magma/det``.
 * Julia and Rust interfaces added, providing a nearly 1-1 correspondence with the C interface, plus some convenience features.
+* New HIP backends for improved tensor basis performance: ``/gpu/hip/shared`` and ``/gpu/hip/gen``.
 
 Performance improvements
 ^^^^^^^^^^^^^^^^^^^^^^^^